python/app-builder/code/chapter11/requirements.txt
Flask-AppBuilder==4.3.0
Flask-Caching==2.0.1
Flask-Compress==1.13
Redis==4.5.4